requirements.txt
matplotlib==3.7.2
numpy==1.24.3
Pillow==9.4.0
Pillow==10.2.0
streamlit==1.29.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
